ref(onboarding): Split sanic onboarding docs
